projects/webapp/requirements.txt
Flask
Werkzeug
mysql-connector-python==8.0.25

